COMPLETE! updated to mod version 1.1.0
Beds will now use every type of plank ItemBlock that is in the
OreDictionary.
Removed excessive recipes, beds are made with any 3 wool blocks and are
seperated by like planks.
Planks that add only an item and no block to the game will not be used.
Beds use the same texture as the plank used to make them for the
underside.
Textures tweaked slightly to be lighter.
Bottom Texture rotated to match plank block.
Shenanigans pulled to drop bed item with correct NBT data, I appologize
to mod authors that use Block.getDrops to get the item dropped from the
bed as I drop it myself with Block.onBlockHarvested.
Fixed Indentation in some classes.