IT FINALLY WORKS!
BED RECIPES WORK ON CLIENT AND SERVER!

Beds should now keep track of the Namespaced name of the plank used for
their frame and their metadata.  They should also update gracefully from
older versions to this new way.

Scissors will show the namespace@metadata used for that block.  remember
this is a debug only item and doesn't have much use!

If a mod that adds planks is REMOVED and there was a bed placed that
used one of those planks, the bed should revert to an Oak Plank frame.

Added a timer for how fast the caching happens in milliseconds, can you
beat 80 milliseconds? :D