Update to 1.8.9
Beds:
SHOULD be the same as be as before.
Made sure to reapply the TE when player sleeps in bed.
Rugs:
Because of how complex they become when using block states, rugs no
longer go up the sides of the block next to them to connect to other
rugs.
Planks:
Added two extra planks, no real way to get them as they're mainly used
for testing the oredict with generating beds.