Finalize work on Eyedropper gui and texture. The eyedropper is not yet functional- it only lets people choose their color right now. Intended functionality is a right click with dyes and water bottles in the inventory (or a crafting operation) creates DYE BOTTLES, which are already implemented. These dyes bottles are then crafted with linen items to dye them.
This is so people can create a multitude of dye sets to keep on hand and try as they please. Or, you know. Open a store. (Is that meta?)

This commit also adds GuiSized- a generic way to have most of the functionality for sized gui windows (like GuiContainer) but without any container code. Use GuiSized.topLeft and GuiSized.guiSize to have fun.