Commit 2.3.4:
Adds a lot of gui stuff. This jar adds the basics of frame editor guis, so peoplecan remove entries and VISUALLY see what's actually whitelisted. This is client-side only at the moment- once we can create the add and editing screens, we can implement packets to sync the changes everywhere.

A single-player tool for now.

Also, this splits meta off into FrameWhitelistEntry so it's abstracted from prying eyes. Get an entry and you can whitelist, blacklist, and get the meta there now.