Moves all references to mod content to ModContent. Creates the bed part system - idea being that a bed has PARTS, and each frame or bed type can accept various pieces to create a customized bed.
Current parts are BLANKETS, SHEETS, DRAWER, and soon, FRAMES.
This is so mods can eventually add completely new bed parts (like pillows, or whatever) and do some fancy things down the line.
Imagine an event a mod could catch where a player sleeps in a BCB bed- they could check if a pillow is on the bed (BedPart.Type.ADDON) and apply a potion effect to the sleeper.